entra: split authentication implementation
EntraAuthentication.cs had grown into a ~900 line monolith mixing
four distinct concerns: core construction/result plumbing, the
interactive/silent public-client user flow, the confidential-client
flows (service principal, managed identity, workload federation), and
MSAL token cache configuration. Reviewing or changing any one concern
required scrolling past the other three.

Make EntraAuthentication a partial class and split it along those
existing seams:

  * EntraAuthentication.cs - constructor, AuthorityIds, and the
    MsalResult wrapper shared by every flow.
  * EntraAuthentication.PublicClient.cs - GetTokenForUserAsync and its
    supporting silent/interactive helpers, flow-type selection, and
    broker/webview/device-code capability detection.
  * EntraAuthentication.ConfidentialClient.cs - service principal,
    managed identity, and workload federation token acquisition, plus
    their confidential client builders and assertion helpers.
  * EntraAuthentication.Caching.cs - MSAL token cache registration and
    the user/app cache property builders.

Also extract the nested MsalHttpClientFactoryAdaptor into its own
top-level internal type in MsalAdaptors.cs, since it has no
dependency on EntraAuthentication's private state and is shared by
both the public- and confidential-client builders.

No method signatures or behavior change here: this is purely moving
existing code into new files along a partial class boundary.

Copilot-Session: 5b041bf4-6167-4ad5-99c1-b622f5a5a5f2
